Refactor Metadata API a little and reformat code
All metadata processing is now done in the metadata package. All code
and tests were updated to reflect this. I'm still
not super happy with the API, but I will wait until the Free refactor to
clean it further.

Fixed a bug where Stage interval metadata was parsed as an Int rather
than a Double.

Updated dependencies and plugins to the latest version.