Major refactor to build an API on Free
LLSM modules have been significantly refactored. The major change is
that the main API is now built using Free monad and the 'onion'
architecture, using the approach described
[here](http://degoes.net/articles/modern-fp-part-2). All logic that
doesn't depend on Free is now in the `core` module to give downstream
developers the choice to use functionality without Free. The Free API
and Algebras are defined in the `api` module.

In terms of LLSM features, the API now allows:

- LLSM metadata to be read and parsed
- LLSM raw data to be read using SCIFIO
- Deskewing
- Writing of LLSM data as HDF5 or OME-TIFF