WIP: OME Metadata Output
This sets out to fix metadata in OMETIFF files and output a companion file that links all individual stacks and provides metadata for the entire dataset. In doing so we have simplified the metadata data types supported for computation, where conversion to SCIFIO metadata types is pushed to the edge and 'raw' parsed metadata is used throughout computation.